refactor: create "state.rs" & "macros.rs"
